Updated post code validation and testing.
COMPASS toolkit has been tested and validated for different distribution and configurations

+++ compass_Qk.py
@@ -31,9 +31,9 @@ def compass_Qk(tIn=None, tParam=None):
     tParam.nx = number of hidden state
     tParam.dLinkMap, the link function between X and continuous variable
     """
-    dLinkMap = int(tParam['dLinkMap'])
+    dLinkMap = tParam['dLinkMap'].astype(int)
     dConstantUpdate = tParam['dConstantUpdate']
-    q_matrix = np.array([np.where(dLinkMap, tIn[k, dLinkMap], 1.0) for k in range(len(tIn))])
+    q_matrix = np.array([np.where(dLinkMap, tIn[k, dLinkMap], 1) for k in range(len(tIn))])
     p_matrix = np.where(dConstantUpdate, 1.0, 0.0)
     return q_matrix, p_matrix
